feat(shared): rewrite models to remove prisma dependency
