+++ libs/constants/src/index.ts
@@ -7,7 +7,6 @@ export * from './consts/steam-default-avatar.const';
 export * from './enums/gamemode.enum';
 export * from './enums/run.enum';
 export * from './enums/role.enum';
-export * from './enums/zone-type.enum';
 export * from './enums/map-status.enum';
 export * from './enums/report-type.enum';
 export * from './enums/activity-type.enum';
@@ -32,9 +31,21 @@ export * from './maps/incompatible-gamemodes.map';
 export * from './maps/map-status-change-rules.map';
 export * from './maps/role-names.map';
 export * from './maps/ban-names.map';
-export * from './types/models';
-export * from './types/queries';
-export * from './types/utils/';
+export * from './types/models/models';
+export * from './types/models/paged-response.model';
+export * from './types/queries/pagination.model';
+export * from './types/queries/map-queries.model';
+export * from './types/queries/run-queries.model';
+export * from './types/queries/user-queries.model';
+export * from './types/queries/admin-queries.model';
+export * from './types/queries/report-queries.model';
+export * from './types/queries/activity-queries.model';
+export * from './types/queries/query.interface';
+export * from './types/queries/order.model';
+export * from './types/utils/flags.type';
+export * from './types/utils/numberify-big-int.type';
+export * from './types/utils/datestring.type';
+export * from './types/utils/vector.type';
 export * from './regexp/youtube-id.regexp';
 export * from './regexp/map-name.regexp';
 export * from './regexp/iso-8601.regexp';
