--- libs/constants/src/types/models/admin/killswitches.model.ts
@@ -1,3 +1,0 @@
-import { KillswitchType } from '../../../enums/killswitch.enum';
-
-export type Killswitches = Record<KillswitchType, boolean>;
