--- libs/constants/src/types/models/auth/jwt-response.model.ts
@@ -1,10 +1,0 @@
-﻿export interface JWTResponseWeb {
-  accessToken: string;
-  refreshToken: string;
-  expiresIn: string;
-}
-
-export interface JWTResponseGame {
-  token: string;
-  length: number;
-}
