--- libs/constants/src/types/models/auth/refresh-token.model.ts
@@ -1,3 +1,0 @@
-﻿export interface RefreshToken {
-  refreshToken: string;
-}
