--- libs/constants/src/types/models/index.ts
@@ -1,46 +1,0 @@
-export * from './map/map.model';
-export * from './map/map-info.model';
-export * from './map/map-image.model';
-export * from './map/map-stats.model';
-export * from './map/map-credit.model';
-export * from './map/map-notify.model';
-export * from './map/map-favorite.model';
-export * from './map/map-submission.model';
-export * from './map/map-submission-approval.model';
-export * from './map/map-submission-dates.model';
-export * from './map/map-submission-placeholder.model';
-export * from './map/map-submission-suggestion.model';
-export * from './map/map-submission-version.model';
-export * from './map/map-review.model';
-export * from './map/map-review-edit.model';
-export * from './map/map-review-comment.model';
-export * from './map/map-review-suggestion.model';
-export * from './map/map-tags.model';
-export * from './map/map-test-invite.model';
-export * from './map/map-zones.model';
-export * from './map/map-list-version.model';
-export * from './map/map-pre-signed-url.model';
-export * from './run/leaderboard.model';
-export * from './run/leaderboard-run.model';
-export * from './run/past-run.model';
-export * from './run/run-session.model';
-export * from './run/completed-run.model';
-export * from './run/leaderboard-stats.model';
-export * from './run/run-session-timestamp.model';
-export * from './auth/jwt-response.model';
-export * from './auth/refresh-token.model';
-export * from './user/user.model';
-export * from './user/follow.model';
-export * from './user/profile.model';
-export * from './user/activity.model';
-export * from './user/user-stats.model';
-export * from './user/notification.model';
-export * from './user/user-maps-summary.model';
-export * from './user/socials.model';
-export * from './admin/admin-activity.model';
-export * from './admin/killswitches.model';
-export * from './socials/twitch-stream.model';
-export * from './stats/base-stats.model';
-export * from './stats/run-stats.model';
-export * from './report/report.model';
-export * from './paged-response.model';
