--- libs/constants/src/types/models/map/map-image.model.ts
@@ -1,16 +1,0 @@
-export interface MapImage {
-  id: string;
-  small: string;
-  medium: string;
-  large: string;
-  xl: string;
-}
-
-export interface UpdateMapImages {
-  imageIDs: string[];
-}
-
-export interface UpdateMapImagesWithFiles {
-  images: File[];
-  data: UpdateMapImages;
-}
