--- libs/constants/src/types/models/map/map-info.model.ts
@@ -1,12 +1,0 @@
-﻿import { MapInfo as PrismaMapInfo } from '@prisma/client';
-
-export type MapInfo = Omit<PrismaMapInfo, 'mapID'>;
-
-export type CreateMapInfo = Pick<
-  MapInfo,
-  'description' | 'youtubeID' | 'creationDate'
->;
-
-export type UpdateMapInfo = Partial<
-  Pick<CreateMapInfo, 'description' | 'youtubeID' | 'creationDate'>
->;
