--- libs/constants/src/types/models/map/map-list-version.model.ts
@@ -1,6 +1,0 @@
-export interface MapListVersion {
-  approved: number;
-  approvedURL: string;
-  submissions: number;
-  submissionsURL: string;
-}
