--- libs/constants/src/types/models/map/map-notify.model.ts
@@ -1,8 +1,0 @@
-﻿import { MapNotify as PrismaMapNotify } from '@prisma/client';
-import { ActivityType } from '../../../enums/activity-type.enum';
-
-export interface MapNotify extends PrismaMapNotify {
-  notifyOn: ActivityType;
-}
-
-export type UpdateMapNotify = Pick<MapNotify, 'notifyOn'>;
