--- libs/constants/src/types/models/map/map-pre-signed-url.model.ts
@@ -1,3 +1,0 @@
-export interface MapPreSignedUrl {
-  url: string;
-}
