--- libs/constants/src/types/models/map/map-review-edit.model.ts
@@ -1,8 +1,0 @@
-import { DateString } from '../../utils';
-
-export interface MapReviewEdit {
-  mainText?: string;
-  resolved?: boolean | null;
-  editorID: number;
-  date: DateString;
-}
