--- libs/constants/src/types/models/map/map-review-suggestion.model.ts
@@ -1,9 +1,0 @@
-import { Gamemode, TrackType } from '../../../';
-
-export interface MapReviewSuggestion {
-  gamemode: Gamemode;
-  trackType: TrackType;
-  trackNum: number;
-  tier?: number;
-  gameplayRating?: number;
-}
