--- libs/constants/src/types/models/map/map-stats.model.ts
@@ -1,8 +1,0 @@
-import { MapStats as PrismaMapStats } from '@prisma/client';
-import { NumberifyBigInt } from '../../utils';
-import { BaseStats } from '../stats/base-stats.model';
-
-export interface MapStats
-  extends NumberifyBigInt<PrismaMapStats, 'timePlayed'> {
-  baseStats: BaseStats;
-}
