--- libs/constants/src/types/models/map/map-submission-dates.model.ts
@@ -1,7 +1,0 @@
-import { DateString } from '../../utils';
-import { MapStatus } from '../../../enums/map-status.enum';
-
-export type MapSubmissionDate = {
-  status: MapStatus;
-  date: DateString;
-};
