--- libs/constants/src/types/models/map/map-submission-placeholder.model.ts
@@ -1,7 +1,0 @@
-import { MapCreditType } from '../../../';
-
-export interface MapSubmissionPlaceholder {
-  alias: string;
-  type: MapCreditType;
-  description?: string;
-}
