--- libs/constants/src/types/models/map/map-submission-suggestion.model.ts
@@ -1,12 +1,0 @@
-import { Gamemode, LeaderboardType, TrackType } from '../../../';
-
-export interface MapSubmissionSuggestion {
-  // TODO: #855 extends JsonObject {
-  trackType: TrackType;
-  trackNum: number;
-  gamemode: Gamemode;
-  tier: number;
-  type: LeaderboardType.RANKED | LeaderboardType.UNRANKED;
-  comment?: string;
-  //  TODO: Tags!
-}
