--- libs/constants/src/types/models/map/map-submission-version.model.ts
@@ -1,19 +1,0 @@
-import { MapSubmissionVersion as PrismaMapSubmissionVersion } from '@prisma/client';
-import { MapZones } from './map-zones.model';
-
-export interface MapSubmissionVersion
-  extends Omit<PrismaMapSubmissionVersion, 'zones'> {
-  zones: MapZones;
-  downloadURL: string;
-  vmfDownloadURL?: string;
-}
-
-export interface CreateMapSubmissionVersion
-  extends Pick<MapSubmissionVersion, 'changelog' | 'zones'> {
-  resetLeaderboards?: boolean;
-}
-
-export interface CreateMapSubmissionVersionWithFiles {
-  vmfs: File[];
-  data: CreateMapSubmissionVersion;
-}
