--- libs/constants/src/types/models/map/map-tags.model.ts
@@ -1,1 +1,0 @@
-export type MapTags = string[];
