--- libs/constants/src/types/models/run/leaderboard-stats.model.ts
@@ -1,6 +1,0 @@
-import { Leaderboard } from './leaderboard.model';
-
-export interface LeaderboardStats {
-  leaderboard: Leaderboard;
-  totalRuns: number;
-}
