--- libs/constants/src/types/models/run/leaderboard.model.ts
@@ -1,6 +1,0 @@
-import { Leaderboard as PrismaLeaderboard } from '@prisma/client';
-import { LeaderboardType } from '../../../enums/leaderboard-type.enum';
-
-export type Leaderboard = PrismaLeaderboard & {
-  type: LeaderboardType;
-};
