--- libs/constants/src/types/models/run/run-session-timestamp.model.ts
@@ -1,4 +1,0 @@
-﻿import { RunSessionTimestamp as PrismaRunSessionTimestamp } from '@prisma/client';
-import { NumberifyBigInt } from '../../utils';
-
-export type RunSessionTimestamp = NumberifyBigInt<PrismaRunSessionTimestamp>;
