--- libs/constants/src/types/models/socials/twitch-stream.model.ts
@@ -1,8 +1,0 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-export interface TwitchStream {
-  title: string;
-  user_name: string;
-  viewer_count: number;
-  started_at: string;
-  thumbnail_url: string;
-}
