--- libs/constants/src/types/models/stats/run-stats.model.ts
@@ -1,6 +1,0 @@
-import { BaseStats } from './base-stats.model';
-
-export interface RunStats {
-  overall: BaseStats;
-  zones?: any; // TODO
-}
