--- libs/constants/src/types/models/user/profile.model.ts
@@ -1,6 +1,0 @@
-import { Profile as PrismaProfile } from '@prisma/client';
-import { Socials } from './socials.model';
-
-export interface Profile extends Omit<PrismaProfile, 'socials'> {
-  socials: Socials;
-}
