--- libs/constants/src/types/models/user/socials.model.ts
@@ -1,14 +1,0 @@
-export type Socials = {
-  Discord?: string;
-  Twitch?: string;
-  YouTube?: string;
-  Github?: string;
-  Twitter?: string;
-  Mastodon?: string;
-  LinkedIn?: string;
-  Instagram?: string;
-  Spotify?: string;
-  Patreon?: string;
-  'Ko-fi'?: string;
-  Paypal?: string;
-};
