--- libs/constants/src/types/models/user/user-maps-summary.model.ts
@@ -1,5 +1,0 @@
-import { MMap } from '@prisma/client';
-
-export interface MapSummary extends Pick<MMap, 'status'> {
-  statusCount: number;
-}
