--- libs/constants/src/types/queries/index.ts
@@ -1,9 +1,0 @@
-export * from './pagination.model';
-export * from './map-queries.model';
-export * from './run-queries.model';
-export * from './user-queries.model';
-export * from './admin-queries.model';
-export * from './report-queries.model';
-export * from './activity-queries.model';
-export * from './query.interface';
-export * from './order.model';
