+++ libs/constants/src/types/queries/map-queries.model.ts
@@ -3,8 +3,22 @@ import { MapStatus } from '../../enums/map-status.enum';
 import { TrackType } from '../../enums/track-type.enum';
 import { Style } from '../../enums/style.enum';
 import { PagedQuery } from './pagination.model';
+import { MapSubmissionType } from '../../enums/map-submission-type.enum';
+import {
+  MapCredit,
+  MapInfo,
+  MapNotify,
+  MapReview,
+  MapReviewComment,
+  MapSubmissionApproval,
+  MapSubmissionPlaceholder,
+  MapSubmissionSuggestion,
+  MapSubmissionVersion,
+  MapZones,
+  MMap
+} from '../../';
 
-//#region Get All
+//#region Map
 
 type BaseMapsGetAllExpand =
   | 'zones'
@@ -69,14 +83,62 @@ export type MapsGetAllUserSubmissionQuery = Omit<
   'submitterID'
 >;
 
-//#endregion
-//#region Get
-
 export type MapsGetExpand = Array<
   MapsGetAllSubmissionExpand[number] | 'submission' | 'testInvites'
 >;
 
 export type MapsGetQuery = { expand?: MapsGetExpand };
+
+export interface CreateMapWithFiles {
+  vmfs: File[];
+  data: CreateMap;
+}
+
+export interface CreateMap extends Pick<MMap, 'name'> {
+  submissionType: MapSubmissionType;
+  suggestions: MapSubmissionSuggestion[];
+  wantsPrivateTesting: boolean;
+  placeholders: MapSubmissionPlaceholder[];
+  testInvites?: number[];
+  info: CreateMapInfo;
+  zones: MapZones;
+  credits: CreateMapCredit[];
+}
+
+export interface UpdateMap
+  extends Partial<
+    Pick<
+      CreateMap,
+      | 'name'
+      | 'suggestions'
+      | 'placeholders'
+      | 'testInvites'
+      | 'credits'
+      | 'submissionType'
+    >
+  > {
+  status?: MapStatus.CONTENT_APPROVAL | MapStatus.FINAL_APPROVAL;
+  info?: UpdateMapInfo;
+  zones?: MapZones;
+  resetLeaderboards?: boolean;
+}
+
+export interface UpdateMapAdmin extends Omit<UpdateMap, 'status'> {
+  status?: MapStatus;
+  finalLeaderboards?: MapSubmissionApproval[];
+}
+
+//#endregion
+//#region Map Info
+
+export type CreateMapInfo = Pick<
+  MapInfo,
+  'description' | 'youtubeID' | 'creationDate'
+>;
+
+export type UpdateMapInfo = Partial<
+  Pick<CreateMapInfo, 'description' | 'youtubeID' | 'creationDate'>
+>;
 
 //#endregion
 //#region Credits
@@ -84,6 +146,28 @@ export type MapsGetQuery = { expand?: MapsGetExpand };
 export type MapCreditsGetExpand = 'user';
 
 export type MapCreditsGetQuery = { expand?: MapCreditsGetExpand };
+
+export type CreateMapCredit = Pick<
+  MapCredit,
+  'userID' | 'type' | 'description'
+>;
+
+//#endregion
+//#region Images
+
+export interface UpdateMapImages {
+  imageIDs: string[];
+}
+
+export interface UpdateMapImagesWithFiles {
+  images: File[];
+  data: UpdateMapImages;
+}
+
+//#endregion
+//#region Notifications
+
+export type UpdateMapNotify = Pick<MapNotify, 'notifyOn'>;
 
 //#endregion
 //#region Runs
@@ -113,6 +197,19 @@ export type MapLeaderboardGetRunQuery = PagedQuery & {
 };
 
 //#endregion
+//#region Submissions
+
+export interface CreateMapSubmissionVersion
+  extends Pick<MapSubmissionVersion, 'changelog' | 'zones'> {
+  resetLeaderboards?: boolean;
+}
+
+export interface CreateMapSubmissionVersionWithFiles {
+  vmfs: File[];
+  data: CreateMapSubmissionVersion;
+}
+
+//#endregion
 //#region Reviews
 
 export type MapReviewsGetExpand = ('map' | 'reviewer' | 'resolver')[];
@@ -126,5 +223,37 @@ export type MapReviewsGetQuery = PagedQuery & {
 export type MapReviewGetIdQuery = {
   expand?: string[];
 };
+
+type PickMapReview = Pick<MapReview, 'mainText'> &
+  Partial<Pick<MapReview, 'suggestions'>>;
+
+export interface CreateMapReview extends PickMapReview {
+  needsResolving?: boolean;
+}
+
+export interface CreateMapReviewWithFiles {
+  images?: File[];
+  data: CreateMapReview;
+}
+
+export interface UpdateMapReview extends Partial<CreateMapReview> {
+  resolved?: boolean | null;
+}
+
+export type AdminUpdateMapReview = Pick<UpdateMapReview, 'resolved'>;
+
+export type CreateMapReviewComment = Pick<MapReviewComment, 'text'>;
+export type UpdateMapReviewComment = CreateMapReviewComment;
+
+//#endregion
+//#region Test Invites
+
+export interface CreateMapTestInvite {
+  userIDs: number[];
+}
+
+export interface UpdateMapTestInvite {
+  accept: boolean;
+}
 
 //#endregion
