--- libs/constants/src/types/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './flags.type';
-export * from './numberify-big-int.type';
-export * from './datestring.type';
-export * from './vector.type';
